+++ src/allotropy/allotrope/schema_mappers/adm/pcr/BENCHLING/_2023/_09/qpcr.py
@@ -121,7 +121,7 @@ class Measurement:
     sample_identifier: str
 
     # Settings
-    pcr_detection_chemistry: str
+    pcr_detection_chemistry: str | None
 
     # Optional measurement metadata
     sample_role_type: str | None = None